edit in reset password form

+++ app/Http/Controllers/ForgotPasswordController.php
@@ -11,7 +11,8 @@ class ForgotPasswordController extends Controller
 {
     public function showForm(Request $request, $token)
     {
-        $email = DB::table('password_resets')->where('token', $request->token)->first()?->email;
+
+        $email = DB::table('password_resets')->where('token', $token)->first()?->email;
 
         if (!$email) {
             return redirect('/');
@@ -24,14 +25,20 @@ class ForgotPasswordController extends Controller
     {
         $email = DB::table('password_resets')->where('token', $request->token)->first()?->email;
         $user = User::where('email', $email)->first();
+        if ($user) {
 
+            if ($request->password == $request->password_confirmation) {
 
-        if ($request->password == $request->password_confirmation) {
-
-            $user->password = Hash::make($request->password_confirmation);
-            $user->save();
-            DB::table('password_resets')->where('token', $request->token)->delete();
+                $user->password = Hash::make($request->password_confirmation);
+                $user->save();
+                DB::table('password_resets')->where('token', $request->token)->delete();
+            }
+            return redirect('/');
         }
-        return redirect('/');
+        else{
+            return response()->json(['errors' => [
+                ['code' => 'not-found', 'message' => __('user not found!')]
+            ]], 404);
+        }
     }
 }
